requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.10.0
azure-search-documents==11.4.0
azure-identity>=1.15.0
qdrant-client>=1.7.0
tenacity>=8.2.0
tiktoken>=0.5.2
python-multipart>=0.0.9
pypdf>=3.17.0
rank_bm25>=0.2.2
sentence-transformers>=2.2.2
jinja2>=3.1.3
ragas>=0.0.22
datasets>=2.16.1

# MLOps
azure-ai-ml>=1.12.0
mlflow>=2.9.0
